src/main/java/com/logmein/jenkins/plugins/pipeline/bamboo/HttpClientFactory.java
/**
 * The MIT License
 *
 * Copyright (c) 2017, LogMeIn, Inc.
 * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
 * associated documentation files (the "Software"), to deal in the Software without restriction, including
 * without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the
 * following conditions:
 *
 * The above copyright notice and this permission notice shall be included in all copies or substantial portions
 * of the Software.  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING
 * BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO
 * EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN
 * ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 * DEALINGS IN THE SOFTWARE.
 */
package com.logmein.jenkins.plugins.pipeline.bamboo;

import org.apache.commons.httpclient.HttpClient;
import org.apache.commons.httpclient.methods.GetMethod;
import org.apache.commons.httpclient.methods.PostMethod;

/**
 * HttpClient factory class to make testing easier.
 *
 * @author Kyle Flavin
 */
public class HttpClientFactory {

    /**
     * Return an HttpClient
     * @return HttpClient object
     */
    public HttpClient getHttpClient() {
        return new HttpClient();
    }

    /**
     * Return a PostMethod
     * @param url POST url
     * @return PostMethod object
     */
    public PostMethod getPostMethod(String url) {
        return new PostMethod(url);
    }

    /**
     * Return a GetMethod
     * @param url GET url
     * @return GetMethod object
     */
    public GetMethod getGetMethod(String url) {
        return new GetMethod(url);
    }
}
